Factor out function resolution code to a separate FunctionResolution class
